+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [0.12.1]
 ### Fixed
 - FlashTP with LAMMPS parallel in torch
-- Single-atom inference failure with ASE+flashTP and LAMMPS ML-IAP
+- Single-atom inference failure with ASE+flashTP, LAMMPS-Torch and LAMMPS ML-IAP
 - CI/CD bugs
 
 ### Added